docker-transcriber/app/requirements.txt
torch
torchaudio
torchlibrosa
pyannote.audio
numpy
ffmpeg-python
whisper-timestamped

